Moves six.py to distarray/externals package.
Makes clear that six is not part of distarray, and allows indirection
down the line to use an already installed six.
