+++ distarray/local/tests/test_base.py
@@ -1,6 +1,6 @@
 import unittest
 
-from distarray.six.moves import range
+from distarray.externals.six.moves import range
 from distarray.local.base import BaseLocalArray
