+++ distarray/local/tests/test_maps.py
@@ -1,7 +1,7 @@
 import unittest
 from distarray.local import maps
 
-from distarray.six.moves import range
+from distarray.externals.six.moves import range
 
 
 class TestNotDistMap(unittest.TestCase):
