+++ distarray/tests/test_distributed_io.py
@@ -11,7 +11,7 @@ import os
 import numpy as np
 from numpy.testing import assert_equal, assert_allclose
 
-from distarray.six.moves import range
+from distarray.externals.six.moves import range
 
 from IPython.parallel import Client
 from distarray.client import DistArray
